You can now select what to download
Created a table that shows all the songs in the playlist and you can uncheck whichever songs you don't want.
This is not yet optimized as you can add the same song or playlist multiple times and it will go through every song regardless of whether it was already downloaded (it won't re-download it but it will still spend time thinking about it).